Create sandbox.html
made a sandbox to test new changes and experiment with html, css, javascript and other HTML data testing. Please use this for testing from now on before committing to the main repo and then when you find something you like in here you can add it to your fashion blog or food blog.